split sketchcomparison to new file; clean up *Result
